fix(schema): align webauthn_credentials user_id type with users.id (BIGINT)
MySQL and PostgreSQL require FK columns to match the referenced column type
exactly. users.id is BIGINT UNSIGNED (MySQL) / BIGINT (PostgreSQL), so
webauthn_credentials.user_id must match. Also aligns the pgsql id column from
SERIAL to BIGINT GENERATED BY DEFAULT AS IDENTITY to match the users table.

Fixes MySQL bootstrap failure: "Referencing column 'user_id' and referenced
column 'id' in foreign key constraint 'fk_wac_user' are incompatible."

+++ Simple-PHP-IPAM/schema.mysql.sql
@@ -644,8 +644,8 @@ CREATE TABLE IF NOT EXISTS backup_history (
 -- webauthn_credentials (v3.15.0 #688: WebAuthn/Passkey credentials)
 -- ---------------------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS webauthn_credentials (
-  id            INT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
-  user_id       INT UNSIGNED NOT NULL,
+  id            BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+  user_id       BIGINT UNSIGNED NOT NULL,
   credential_id VARBINARY(255) NOT NULL, -- PARAM_LOB required on all writes: see ipam_bind_binary()
   public_key    TEXT NOT NULL,
   sign_count    INT UNSIGNED NOT NULL DEFAULT 0,
